Removed last remaining warnings
