+++ Aggregator.ConsoleApp/GlobalSuppressions.cs
@@ -1,0 +1,5 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Maintainability", "S3220:Method calls should not resolve ambiguously to overloads with \"params\"", Justification = "<Pending>", Scope = "member", Target = "~M:Aggregator.ConsoleApp.ConsoleTextLogger.Log(Aggregator.Core.Monitoring.LogLevel,System.String,System.Object[])")]
