+++ Aggregator.Core/Aggregator.Core.csproj
@@ -111,10 +111,16 @@
     <Compile Include="Configuration\ScopeMatchResult.cs" />
     <Compile Include="Configuration\TemplateScope.cs" />
     <Compile Include="Configuration\WorkItemTypeScope.cs" />
+    <Compile Include="Extensions\BaseFieldValueValidator.cs" />
     <Compile Include="Extensions\DoubleFixFieldDecorator.cs" />
     <Compile Include="Extensions\FieldValueValidationDecorator.cs" />
+    <Compile Include="Extensions\IncorrectDataTypeFieldValidator.cs" />
+    <Compile Include="Extensions\InvalidValueFieldValueValidator.cs" />
+    <Compile Include="Extensions\NullAssignmentToRequiredFieldValueValidator.cs" />
+    <Compile Include="Extensions\ValueAssignmentToReadonlyFieldValueValidator.cs" />
     <Compile Include="Facade\ProcessTemplateVersionWrapper.cs" />
     <Compile Include="Facade\RevisionWrapper.cs" />
+    <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="Interfaces\IFieldExposed.cs" />
     <Compile Include="Interfaces\IProcessTemplateVersion.cs" />
     <Compile Include="Interfaces\IRevision.cs" />
