+++ Aggregator.Core/Interfaces/IWorkItemExposed.cs
@@ -42,7 +42,9 @@ namespace Aggregator.Core.Interfaces
         // links management
         void AddWorkItemLink(IWorkItemExposed destination, string linkTypeName);
 
-        void AddHyperlink(string destination, string comment = "");
+        void AddHyperlink(string destination);
+
+        void AddHyperlink(string destination, string message);
 
         // state helpers; must be on interface to work on WorkItemLazyReference
         void TransitionToState(string state, string comment);
