+++ UnitTests.Core/Mock/WorkItemMock.cs
@@ -157,7 +157,12 @@ namespace UnitTests.Core.Mock
             }
         }
 
-        public void AddHyperlink(string destination, string comment = "")
+        public void AddHyperlink(string destination)
+        {
+            this.AddHyperlink(destination, string.Empty);
+        }
+
+        public void AddHyperlink(string destination, string comment)
         {
             throw new NotImplementedException();
         }
